refactor(commands/key_packages): move commands to individual files
