+++ src-tauri/src/commands/key_packages/mod.rs
@@ -1,0 +1,7 @@
+mod delete_all_key_packages;
+mod publish_new_key_package;
+mod valid_key_package_exists_for_user;
+
+pub use delete_all_key_packages::delete_all_key_packages;
+pub use publish_new_key_package::publish_new_key_package;
+pub use valid_key_package_exists_for_user::valid_key_package_exists_for_user;
